700-750ms now

+++ get_random_image.py
@@ -6,7 +6,8 @@ req_height = 1080
 extension = '.jpg'
 dir_path = pathlib.Path('sample-inputs')
 url = f'https://picsum.photos/{req_width}/{req_height}{extension}'
-for i in range(20):
+for i in range(100):
     path = dir_path / f'lorem-picsum-{i}{extension}'
-    print(f'Requesting image from {url}')
-    urllib.request.urlretrieve(url, path)
+    if not path.exists():
+        print(f'Requesting image from {url}')
+        urllib.request.urlretrieve(url, path)
